🐛 소셜로그인 시 일반로그인처럼 알럿작동되는 문제 해결
